update pages after fixing error
