+++ src/context/ProductContext.jsx
@@ -1,6 +1,6 @@
 import React, { createContext, useState, useEffect } from "react";
 
-// context
+// create context
 
 export const ProductContext = createContext();
